:sparkles: Add profile image upload functionality
Profile image selection added to member creation form. Known issue: if image is not selected an error is throw, planned solution is to alert users if profile image has not been selected and if they choose to skip the profile picture they will be assigned a random avatar image on form submission.